backend/requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI Agent
openai==1.3.0
google-generativeai==0.7.2
anthropic==0.39.0

# Database Drivers
pymongo==4.6.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
aiosqlite==0.19.0

# AWS
boto3==1.29.0
botocore==1.32.0

# Schema & Validation
pydantic==2.5.0
pydantic-settings==2.0.0
jsonschema==4.20.0

# Visualization
graphviz==0.20.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
httpx==0.25.1
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
